Finalizando nova implementação de home e pagina de equipe
